requirements.txt
discord.py
Pillow
requests
python-dotenv
aiohttp
aiomysql
flask   